change depracated np.float type hints to float
Former-commit-id: 1c8720eeb8310e9abf85de15831d5edf8a95d8e9

+++ scripts/compute_cell_area.py
@@ -156,7 +156,7 @@ def threshold_images(
 
 def compute_areas(
     imgs: List[npt.NDArray], circ_pix_area: List[int]
-) -> npt.NDArray[np.float_]:
+) -> npt.NDArray[float]:
     """Compute non-zero pixel area of thresholded images.
 
     Args:
